# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SpiderX is a private tracker server.",
        "title": "SpiderX API",
        "contact": {
            "name": "EndPot",
            "url": "https://github.com/endpot"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
        "/auth.getMine": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "获取个人信息",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.Mine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auth.login": {
            "post": {
                "description": "用户使用用户名密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录参数",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "token": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auth.logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户退出登录态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登出",
                "responses": {
                    "204": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auth.refreshToken": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "刷新登录密钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "刷新登录密钥",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "token": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auth.register": {
            "post": {
                "description": "用户自主注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册参数",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auth.reset": {
            "post": {
                "description": "重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "重置密码参数",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ResetRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auth.sendCode": {
            "post": {
                "description": "发送校验码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "发送校验码",
                "parameters": [
                    {
                        "description": "发送校验码参数",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SendCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "display_name",
                "password"
            ],
            "properties": {
                "display_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "spider"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "123456789"
                }
            }
        },
        "auth.Mine": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像地址",
                    "type": "string",
                    "example": "https://i.endpot.com/x/xxx.jpg"
                },
                "display_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "spider"
                },
                "down_traffic": {
                    "description": "下载量",
                    "type": "integer",
                    "example": 1
                },
                "email": {
                    "description": "用户邮箱",
                    "type": "string",
                    "example": "spider@spider.com"
                },
                "grade": {
                    "description": "用户等级",
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "roles": {
                    "description": "用户角色",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "up_traffic": {
                    "description": "上传量",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "auth.RegisterRequest": {
            "type": "object",
            "required": [
                "display_name",
                "email",
                "password",
                "password_confirm"
            ],
            "properties": {
                "display_name": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "spider"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "spider@spider.com"
                },
                "invite_code": {
                    "description": "邀请码",
                    "type": "string",
                    "example": "spider"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8,
                    "example": "123456789"
                },
                "password_confirm": {
                    "description": "密码确认",
                    "type": "string",
                    "example": "123456789"
                },
                "verify_code": {
                    "description": "校验码",
                    "type": "string",
                    "example": "spider"
                }
            }
        },
        "auth.ResetRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "password_confirm"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "spider@spider.com"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8,
                    "example": "123456789"
                },
                "password_confirm": {
                    "description": "密码确认",
                    "type": "string",
                    "example": "123456789"
                },
                "verify_code": {
                    "description": "校验码",
                    "type": "string",
                    "example": "spider"
                }
            }
        },
        "auth.SendCodeRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "用户邮箱",
                    "type": "string",
                    "example": "spider@spider.com"
                }
            }
        },
...
        "response.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}